+++ nft-studio/src/components/Root.tsx
@@ -1,10 +1,8 @@
 import { GlobalStyle as FabricGlobalStyle } from '@centrifuge/fabric'
-import centrifugeLight from '@centrifuge/fabric/dist/theme/centrifugeLight'
 import { OverlayProvider } from '@react-aria/overlays'
 import * as React from 'react'
 import { QueryClient, QueryClientProvider } from 'react-query'
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
-import { DefaultTheme, ThemeProvider } from 'styled-components'
 import { AccountNFTsPage } from '../pages/AccountNFTs'
 import { CollectionPage } from '../pages/Collection'
 import { CollectionsPage } from '../pages/Collections'
@@ -29,30 +27,6 @@ import { TransactionProvider } from './TransactionsProvider'
 import { TransactionToasts } from './TransactionToasts'
 import { Web3Provider } from './Web3Provider'
 
-const darkTheme: DefaultTheme = {
-  ...centrifugeLight,
-  sizes: {
-    ...centrifugeLight.sizes,
-    container: '100%',
-    navBarHeight: 72,
-    navBarHeightMobile: 64,
-    dialog: 564,
-  },
-  colors: {
-    ...centrifugeLight.colors,
-    placeholderBackground: centrifugeLight.colors.backgroundSecondary,
-  },
-  typography: {
-    ...centrifugeLight.typography,
-    headingLarge: {
-      fontSize: [24, 24, 36],
-      lineHeight: 1.25,
-      fontWeight: 600,
-      color: 'textPrimary',
-    },
-  },
-}
-
 const queryClient = new QueryClient({
   defaultOptions: {
     queries: {
@@ -65,107 +39,105 @@ export const Root: React.FC = () => {
   const showOnlyNFT = useDebugFlags().showOnlyNFT
   return (
     <QueryClientProvider client={queryClient}>
-      <ThemeProvider theme={darkTheme}>
-        <GlobalStyle />
-        <FabricGlobalStyle />
-        <OverlayProvider>
-          <HostPermissionsProvider>
-            <Web3Provider>
-              <CentrifugeProvider>
-                <TransactionProvider>
-                  <TransactionToasts />
-                  <Router>
-                    <LoadBoundary>
-                      {showOnlyNFT ? (
-                        <Switch>
-                          <Route path="/collection/:cid/object/mint">
-                            <MintNFTPage />
-                          </Route>
-                          <Route path="/collection/:cid/object/:nftid/new-asset">
-                            <CreateLoanPage />
-                          </Route>
-                          <Route path="/collection/:cid/object/:nftid">
-                            <NFTPage />
-                          </Route>
-                          <Route path="/collection/:cid">
-                            <CollectionPage />
-                          </Route>
-                          <Route path="/account">
-                            <AccountNFTsPage />
-                          </Route>
-                          <Route path="/nfts">
-                            <CollectionsPage />
-                          </Route>
-                          <Route exact path="/">
-                            <CollectionsPage />
-                          </Route>
-                          <Route>
-                            <NotFoundPage />
-                          </Route>
-                        </Switch>
-                      ) : (
-                        <Switch>
-                          <Route path="/collection/:cid/object/mint">
-                            <MintNFTPage />
-                          </Route>
-                          <Route path="/collection/:cid/object/:nftid/new-asset">
-                            <CreateLoanPage />
-                          </Route>
-                          <Route path="/collection/:cid/object/:nftid">
-                            <NFTPage />
-                          </Route>
-                          <Route path="/collection/:cid">
-                            <CollectionPage />
-                          </Route>
-                          <Route path="/account">
-                            <AccountNFTsPage />
-                          </Route>
-                          <Route path="/nfts">
-                            <CollectionsPage />
-                          </Route>
-                          <Route path="/pools/:pid/assets/:aid">
-                            <LoanPage />
-                          </Route>
-                          <Route path="/pools/:pid">
-                            <PoolPage />
-                          </Route>
-                          <Route path="/pools">
-                            <PoolsPage />
-                          </Route>
-                          <Route path="/pool/new">
-                            <PoolFormPage />
-                          </Route>
-                          <Route path="/issuers/assets">
-                            <LoansPage />
-                          </Route>
-                          <Route path="/investments/tokens/:pid/:tid">
-                            <TokenPage />
-                          </Route>
-                          <Route path="/investments/tokens">
-                            <TokensPage />
-                          </Route>
-                          <Route path="/investments">
-                            <TokensPage />
-                          </Route>
-                          <Route path="/issuers/managed-pools">
-                            <ManagedPoolsPage />
-                          </Route>
-                          <Route exact path="/">
-                            <PoolsPage />
-                          </Route>
-                          <Route>
-                            <NotFoundPage />
-                          </Route>
-                        </Switch>
-                      )}
-                    </LoadBoundary>
-                  </Router>
-                </TransactionProvider>
-              </CentrifugeProvider>
-            </Web3Provider>
-          </HostPermissionsProvider>
-        </OverlayProvider>
-      </ThemeProvider>
+      <GlobalStyle />
+      <FabricGlobalStyle />
+      <OverlayProvider>
+        <HostPermissionsProvider>
+          <Web3Provider>
+            <CentrifugeProvider>
+              <TransactionProvider>
+                <TransactionToasts />
+                <Router>
+                  <LoadBoundary>
+                    {showOnlyNFT ? (
+                      <Switch>
+                        <Route path="/collection/:cid/object/mint">
+                          <MintNFTPage />
+                        </Route>
+                        <Route path="/collection/:cid/object/:nftid/new-asset">
+                          <CreateLoanPage />
+                        </Route>
+                        <Route path="/collection/:cid/object/:nftid">
+                          <NFTPage />
+                        </Route>
+                        <Route path="/collection/:cid">
+                          <CollectionPage />
+                        </Route>
+                        <Route path="/account">
+                          <AccountNFTsPage />
+                        </Route>
+                        <Route path="/nfts">
+                          <CollectionsPage />
+                        </Route>
+                        <Route exact path="/">
+                          <CollectionsPage />
+                        </Route>
+                        <Route>
+                          <NotFoundPage />
+                        </Route>
+                      </Switch>
+                    ) : (
+                      <Switch>
+                        <Route path="/collection/:cid/object/mint">
+                          <MintNFTPage />
+                        </Route>
+                        <Route path="/collection/:cid/object/:nftid/new-asset">
+                          <CreateLoanPage />
+                        </Route>
+                        <Route path="/collection/:cid/object/:nftid">
+                          <NFTPage />
+                        </Route>
+                        <Route path="/collection/:cid">
+                          <CollectionPage />
+                        </Route>
+                        <Route path="/account">
+                          <AccountNFTsPage />
+                        </Route>
+                        <Route path="/nfts">
+                          <CollectionsPage />
+                        </Route>
+                        <Route path="/pools/:pid/assets/:aid">
+                          <LoanPage />
+                        </Route>
+                        <Route path="/pools/:pid">
+                          <PoolPage />
+                        </Route>
+                        <Route path="/pools">
+                          <PoolsPage />
+                        </Route>
+                        <Route path="/pool/new">
+                          <PoolFormPage />
+                        </Route>
+                        <Route path="/issuers/assets">
+                          <LoansPage />
+                        </Route>
+                        <Route path="/investments/tokens/:pid/:tid">
+                          <TokenPage />
+                        </Route>
+                        <Route path="/investments/tokens">
+                          <TokensPage />
+                        </Route>
+                        <Route path="/investments">
+                          <TokensPage />
+                        </Route>
+                        <Route path="/issuers/managed-pools">
+                          <ManagedPoolsPage />
+                        </Route>
+                        <Route exact path="/">
+                          <PoolsPage />
+                        </Route>
+                        <Route>
+                          <NotFoundPage />
+                        </Route>
+                      </Switch>
+                    )}
+                  </LoadBoundary>
+                </Router>
+              </TransactionProvider>
+            </CentrifugeProvider>
+          </Web3Provider>
+        </HostPermissionsProvider>
+      </OverlayProvider>
     </QueryClientProvider>
   )
 }
